account checker features for signup page

+++ src/js/signUp.js
@@ -9,6 +9,20 @@ const passwordSignal = form.querySelector("[data-signal='password-form-signal']"
 const submitBtn = document.querySelector(".submit-btn");
 
 
+// ACCOUNT CHECKER
+async function isAccountAlreadyExist(email) {
+    const response = await axios.get(`/data/${email}`);
+    const {data} = await response;
+    const {users} = data;
+
+    for(let x = 0; x < users.length; x++) {
+        if(users[x].email === email) {
+            return true;
+        }
+    }
+    return false;
+}
+
 // Form Events
 form.addEventListener("submit", async (event) => {
     event.preventDefault();
@@ -20,22 +34,30 @@ form.addEventListener("submit", async (event) => {
         const password = passwordIn.value;
         const postData = {username, email, password};
 
-        try {
-            const {data} = await axios.post("/register", postData);
-
-            if(data.success) {
-                const localStorage = window.localStorage
-                localStorage.setItem("colorant_user", JSON.stringify(data.formData));
-
-                // Change path to success state
-                window.location.href = "../"
+        // CHECK IF ACCOUNT IS ALREADY REGISTERED
+        if(await isAccountAlreadyExist(email)) {
+            emailSignal.classList.add("fail");
+            emailSignal.querySelector("span").textContent = "email is already registered";
+            return;
+        }else {
+            try {
+                const {data} = await axios.post("/register", postData);
+    
+                if(data.success) {
+                    const localStorage = window.localStorage
+                    localStorage.setItem("colorant_user", JSON.stringify(data.formData));
+    
+                    // Redirect to homepage with login state
+                    window.location.href = "../"
+                }
+    
+            } catch (error) {
+                console.log(error);
             }
-
-        } catch (error) {
-            console.log(error.response.data.msg);
         }
-    }
+        
 
+    }
 })
 
 // Form validator events
@@ -46,16 +68,16 @@ usernameIn.addEventListener("input", event => {
 
     if(val === "") {
         usernameSignal.classList.add("fail");
-        return validInfo.innerText = "please fill in your password";
+        return validInfo.textContent = "please fill in your password";
     } else {
         let error = "";
 
         // validation test
         if(!/^[a-zA-Z]+.+$/.test(val)) {
             error = "Username must start with letter";
-        } else if(!/^[\w\W\d]{6,}$/.test(val)) {
-            error = "Username must contains 6+ characters";
-        } else if(!/^[\w\W\d]{6,16}$/.test(val)) {
+        } else if(!/^[\w\W\d]{3,}$/.test(val)) {
+            error = "Username must contains 3 or more characters";
+        } else if(!/^[\w\W\d]{3,16}$/.test(val)) {
             error = "Username must be less than 16 characters";
         } else if(!/^[\w\d]+$/.test(val)) {
             error = "Username must not contains any symbol or special characters";
@@ -63,10 +85,10 @@ usernameIn.addEventListener("input", event => {
 
         if(error) {
             usernameSignal.classList.add("fail");
-            return validInfo.innerText = error;
+            return validInfo.textContent = error;
         } else {
             usernameSignal.classList.remove("fail");
-            return validInfo.innerText = "";
+            return validInfo.textContent = "";
         }
     }
 });
@@ -78,21 +100,21 @@ emailIn.addEventListener("input", event => {
 
     if(val === "") {
         emailSignal.classList.add("fail");
-        return validInfo.innerText = "please fill in your password";
+        return validInfo.textContent = "please fill in your password";
     } else {
         let error = "";
 
         // validation test
-        if(!/^([\w\d]{3,50})@([a-z]{3,}).([a-z]{2,})$/.test(val)) {
+        if(!/^([\w\d]{3,50})@([a-z-]{3,})[.]{1,}([a-z]{2,})$/.test(val)) {
             error = "Please fill in the email field correctly";
         }
 
         if(error) {
             emailSignal.classList.add("fail");
-            return validInfo.innerText = error;
+            return validInfo.textContent = error;
         } else {
             emailSignal.classList.remove("fail");
-            return validInfo.innerText = "";
+            return validInfo.textContent = "";
         }
     }
 });
@@ -104,7 +126,7 @@ passwordIn.addEventListener("input", event => {
 
     if(val === "") {
         passwordSignal.classList.add("fail");
-        return validInfo.innerText = "please fill in your password";
+        return validInfo.textContent = "please fill in your password";
     } else {
         let error = "";
         if(!/^[\w\W\d]{6,}$/.test(val)) {
@@ -121,10 +143,10 @@ passwordIn.addEventListener("input", event => {
 
         if(error) {
             passwordSignal.classList.add("fail");
-            return validInfo.innerText = error;
+            return validInfo.textContent = error;
         } else {
             passwordSignal.classList.remove("fail");
-            return validInfo.innerText = "";
+            return validInfo.textContent = "";
         }
     }
